+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.1.0
+
+- Added browser-compatible version as `browser_resource.dart` library.
+  Only needed because configurable imports are not available yet.
+
 ## 1.0.0
 
 - Initial version